remove old Network Manager
