+++ geoplateforme/toolbelt/__init__.py
@@ -1,3 +1,4 @@
 #! python3  # noqa: E265
 from .log_handler import PlgLogger  # noqa: F401
+from .network_manager import NetworkRequestsManager  # noqa: F401
 from .preferences import PlgOptionsManager  # noqa: F401
